infra: remove obsolete scripts, move interface to core
