+++ CausalEstimate/__init__.py
@@ -1,3 +1,3 @@
-from CausalEstimate.interface.multi_estimator import MultiEstimator
+from CausalEstimate.core.multi_estimator import MultiEstimator
 
 __all__ = ["MultiEstimator"]
